+++ classification_3D/preprocess.py
@@ -3,15 +3,15 @@ import math
 import numpy as np
 import pandas as pd
 
-from intracranial_hemorrhage_detection.classification_3D.params import input_slice_size, max_slices_per_study
+from intracranial_hemorrhage_detection.classification_3D.params import dtype, input_slice_size, max_slices_per_study
 from intracranial_hemorrhage_detection.constants import folder_path
 
 def preprocess_voxels(voxels):
     "Inputs a list of voxel. Outputs a numpy array with the right shape"
     number_of_slices = len(voxels)
-    voxels = [np.array(pixels, dtype=np.float32) for pixels in voxels]
+    voxels = [np.array(pixels, dtype=np.float32) for pixels in voxels] # must use float32 for cv2.resize, even if we use another dtype later
     voxels= [cv2.resize(pixels, (input_slice_size, input_slice_size)) for pixels in voxels]
-    voxels = np.array(voxels, dtype=np.float32)
+    voxels = np.array(voxels, dtype=dtype)
     voxels -= np.min(voxels)
     voxels /= np.max(voxels)
     voxels = np.reshape(voxels, (number_of_slices, input_slice_size, input_slice_size, 1))
