+++ classification_3D/training_generator.py
@@ -2,7 +2,7 @@ import numpy as np
 import random
 
 from intracranial_hemorrhage_detection.classification_3D.generate_3D_image import generate_3D_image
-from intracranial_hemorrhage_detection.classification_3D.params import batch_size
+from intracranial_hemorrhage_detection.classification_3D.params import batch_size, dtype
 from intracranial_hemorrhage_detection.classification_3D.preprocess import add_padding, get_study_wise_data
 from intracranial_hemorrhage_detection.preprocess import get_all_true_labels
 
@@ -20,7 +20,7 @@ def training_generator():
         grouped_slice_ids = grouped_slice_ids[batch_size:]
 
         X = [generate_3D_image(slice_ids) for slice_ids in batch]
-        Y = [np.array([true_labels[id] for id in slice_ids], dtype=np.float32) for slice_ids in batch]
-        X = np.array([add_padding(x) for x in X])
-        Y = np.array([add_padding(y) for y in Y])
+        Y = [np.array([true_labels[id] for id in slice_ids], dtype=dtype) for slice_ids in batch]
+        X = np.array([add_padding(x) for x in X], dtype=dtype)
+        Y = np.array([add_padding(y) for y in Y], dtype=dtype)
         yield ([X], [Y])
